Improve setup.py.

+++ setup.py
@@ -20,7 +20,6 @@ def call_command(cmd):
 
 def get_version(compiler):
     ver = call_command('gfortran -dumpversion')
-    print(ver)
     if ver and '.' not in ver:
         ver = call_command(compiler + ' -dumpfullversion')
     return ver
@@ -59,6 +58,46 @@ def make_library():
         raise IOError('gcc must be al least version 4.8.5.')
     
     if platform.system() == 'Windows':
+        subprocess.call(('del ./calc11/lib/*.so'), shell=True)
+        subprocess.call(('del ./calc11/src/*.o'), shell=True)
+        subprocess.call(('del ./calc11/src/*.so'), shell=True)
+    else:
+        subprocess.call(('rm -rf ./calc11/lib/*.so'), shell=True)
+        subprocess.call(('rm -rf ./calc11/src/*.o'), shell=True)
+        subprocess.call(('rm -rf ./calc11/src/*.so'), shell=True)
+        
+    os.chdir('./calc11/src')
+    
+    cmd  = 'gfortran -ffree-form -ffree-line-length-none -fPIC -I. -g -O2 -c almacalc.f  cut1m.f  cwobm.f  ainit.f  csitm.f  adrvr.f  cctiu.f  catiu.f  cdtdb.f  cpepu.f  cnutm.f  cnutu.f  cnutu6.f  cdiuu.f  cm20u.f  crosu.f  cetdm.f  cptdm.f  cocem.f  hardisp.f  dsitu.f  astrm.f  catmm.f  caxom.f  cuvm.f  ctheu.f  cmatu.f  cmabd.f  cvecu.f  dkill.f calc11_wrapper.f90'
+    #print(cmd)
+    if subprocess.call(cmd, shell=True) != 0:
+        raise IOError('Compilation failed')
+                    
+    cmd = 'gcc -fPIC -c almaout.c '
+    #print(cmd)
+    if subprocess.call(cmd, shell=True) != 0:
+        raise IOError('Compilation failed')
+    
+    cmd = 'gfortran -shared -o libcalc11.so almacalc.o  cut1m.o  cwobm.o  ainit.o  csitm.o  adrvr.o  cctiu.o  catiu.o  cdtdb.o  cpepu.o  cnutm.o  cnutu.o  cnutu6.o  cdiuu.o  cm20u.o  crosu.o  cetdm.o  cptdm.o  cocem.o  hardisp.o  dsitu.o  astrm.o  catmm.o  caxom.o  cuvm.o  ctheu.o  cmatu.o  cmabd.o  cvecu.o  dkill.o  almaout.o calc11_wrapper.o'
+    #print(cmd)
+    if subprocess.call(cmd, shell=True) != 0:
+        raise IOError('Compilation failed')
+    
+    
+    if platform.system() == 'Windows':
+        subprocess.call(('move libcalc11.so ../lib'), shell=True)
+    else:
+        subprocess.call(('mv libcalc11.so ../lib'), shell=True)
+    os.chdir('../../')
+    
+    '''
+    if get_version('gcc') is None:
+        raise IOError('gcc version 4.8.5 or later is required.')
+    
+    if compare_version(get_version('gcc'),'4.8.5') < 0:
+        raise IOError('gcc must be al least version 4.8.5.')
+    
+    if platform.system() == 'Windows':
         subprocess.call(('rm -rf calc11\lib\*.so'), shell=True)
         os.chdir('.\calc11\src')
         subprocess.call(('del -rf *.o','del -rf *.so'), shell=True)
@@ -72,7 +111,7 @@ def make_library():
     
     SOURCES_F = ['almacalc', 'cut1m', 'cwobm', 'ainit', 'csitm', 'adrvr', 'cctiu', 'catiu', 'cdtdb', 'cpepu', 'cnutm', 'cnutu', 'cnutu6', 'cdiuu', 'cm20u', 'crosu', 'cetdm', 'cptdm', 'cocem', 'hardisp', 'dsitu', 'astrm', 'catmm', 'caxom', 'cuvm', 'ctheu', 'cmatu', 'cmabd', 'cvecu', 'dkill']
     cmd  = COMPILER_F + ' ' + FLAGS_F + ' ' + ' '.join([ s + '.f ' for s in SOURCES_F])
-    os.system('pwd')
+    print(cmd)
     if subprocess.call(cmd, shell=True) != 0:
         raise IOError('Compilation failed')
                     
@@ -80,16 +119,19 @@ def make_library():
     FLAGS_C = '-fPIC -c'
     SOURCES_C = ['almaout']
     cmd = COMPILER_C + ' ' + FLAGS_C + ' '  + ' '.join([s + '.c ' for s in SOURCES_C])
+    print(cmd)
     if subprocess.call(cmd, shell=True) != 0:
         raise IOError('Compilation failed')
     
     SOURCES = SOURCES_F + SOURCES_C
     FLAGS = '-shared -o'
     cmd = COMPILER_F + ' ' + FLAGS + ' ' + DLLNAME + ' ' + ' '.join([ s + '.o ' for s in SOURCES])
+    print(cmd)
     if subprocess.call(cmd, shell=True) != 0:
         raise IOError('Compilation failed')
                     
     cmd = COMPILER_F + ' -fPIC -shared -o calc11_wrapper.so calc11_wrapper.f90 -L. -lcalc11'
+    print(cmd)
     if subprocess.call(cmd, shell=True) != 0:
         raise IOError('Compilation failed')
     
@@ -102,6 +144,7 @@ def make_library():
         subprocess.call(('mv libcalc11.so ../lib'), shell=True)
     
     os.chdir('../../')
+    '''
     
     
 #    cmd = 'make -C ./calc11/src/ clean'
